Fix linting

+++ docker-entrypoint.sh
@@ -63,7 +63,7 @@ else
     wget https://raw.githubusercontent.com/bnb-chain/bsc-snapshots/main/dist/fetch-snapshot.sh
     bash fetch-snapshot.sh -d -D "${__snap_dir}" "${SNAPSHOT}"
     bash fetch-snapshot.sh -e -D "${__snap_dir}" -E "${__data_dir}" "${SNAPSHOT}"
-    rm -rf "${__snap_dir}/"
+    rm -rf "${__snap_dir:?}/"
 
     # try to find the directory
     __search_dir="geth/chaindata"
